# frontend/src/lib/api/openapi.json
{
	"components": {
		"schemas": {
...
			"HTTPValidationError": {
				"properties": {
					"detail": {
						"items": {
							"$ref": "#/components/schemas/ValidationError"
						},
						"title": "Detail",
						"type": "array"
					}
				},
				"title": "HTTPValidationError",
				"type": "object"
			},
...
			"ValidationError": {
				"properties": {
					"ctx": {
						"title": "Context",
						"type": "object"
					},
					"input": {
						"title": "Input"
					},
					"loc": {
						"items": {
							"anyOf": [
								{
									"type": "string"
								},
								{
									"type": "integer"
								}
							]
						},
						"title": "Location",
						"type": "array"
					},
					"msg": {
						"title": "Message",
						"type": "string"
					},
					"type": {
						"title": "Error Type",
						"type": "string"
					}
				},
				"required": [
					"loc",
					"msg",
					"type"
				],
				"title": "ValidationError",
				"type": "object"
			},
...
		}
	},
	"info": {
		"description": "Self-hosted reader for collections of PDF periodicals.",
		"title": "Paperstand",
		"version": "0.1.0"
	},
	"openapi": "3.1.0",
	"paths": {
...
		"/api/issues/{issue_id}/cover.jpg": {
			"get": {
				"description": "The issue's cover, rendered on demand when the cache has lost it.",
				"operationId": "issue_cover",
				"parameters": [
					{
						"in": "path",
						"name": "issue_id",
						"required": true,
						"schema": {
							"title": "Issue Id",
							"type": "string"
						}
					},
					{
						"description": "Cache buster: the PDF's mtime",
						"in": "query",
						"name": "v",
						"required": false,
						"schema": {
							"anyOf": [
								{
									"type": "string"
								},
								{
									"type": "null"
								}
							],
							"description": "Cache buster: the PDF's mtime",
							"title": "V"
						}
					}
				],
				"responses": {
					"200": {
						"content": {
							"image/jpeg": {}
						},
						"description": "The image"
					},
					"304": {
						"description": "The client's copy is current"
					},
					"422": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						},
						"description": "Validation Error"
					}
				},
				"summary": "Cover",
				"tags": [
					"images"
				]
			},
			"head": {
				"description": "The issue's cover, rendered on demand when the cache has lost it.",
				"operationId": "issue_cover_head",
				"parameters": [
					{
						"in": "path",
						"name": "issue_id",
						"required": true,
						"schema": {
							"title": "Issue Id",
							"type": "string"
						}
					},
					{
						"description": "Cache buster: the PDF's mtime",
						"in": "query",
						"name": "v",
						"required": false,
						"schema": {
							"anyOf": [
								{
									"type": "string"
								},
								{
									"type": "null"
								}
							],
							"description": "Cache buster: the PDF's mtime",
							"title": "V"
						}
					}
				],
				"responses": {
					"200": {
						"content": {
							"image/jpeg": {}
						},
						"description": "The image"
					},
					"304": {
						"description": "The client's copy is current"
					},
					"422": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						},
						"description": "Validation Error"
					}
				},
				"summary": "Cover",
				"tags": [
					"images"
				]
			}
		},
...
		"/api/issues/{issue_id}/pages/{page}.webp": {
			"get": {
				"description": "One page of an issue as a WebP image, rendered on demand.",
				"operationId": "issue_page",
				"parameters": [
					{
						"in": "path",
						"name": "issue_id",
						"required": true,
						"schema": {
							"title": "Issue Id",
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "page",
						"required": true,
						"schema": {
							"title": "Page",
							"type": "integer"
						}
					},
					{
						"description": "Snapped to the next width up",
						"in": "query",
						"name": "w",
						"required": false,
						"schema": {
							"anyOf": [
								{
									"maximum": 100000,
									"minimum": 1,
									"type": "integer"
								},
								{
									"type": "null"
								}
							],
							"description": "Snapped to the next width up",
							"title": "W"
						}
					},
					{
						"description": "Cache buster: the PDF's mtime",
						"in": "query",
						"name": "v",
						"required": false,
						"schema": {
							"anyOf": [
								{
									"type": "string"
								},
								{
									"type": "null"
								}
							],
							"description": "Cache buster: the PDF's mtime",
							"title": "V"
						}
					}
				],
				"responses": {
					"200": {
						"content": {
							"image/webp": {}
						},
						"description": "The rendered page"
					},
					"304": {
						"description": "The client's copy is current"
					},
					"404": {
						"description": "No such issue, or no such page"
					},
					"422": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						},
						"description": "Validation Error"
					}
				},
				"summary": "Page Image",
				"tags": [
					"images"
				]
			},
			"head": {
				"description": "One page of an issue as a WebP image, rendered on demand.",
				"operationId": "issue_page_head",
				"parameters": [
					{
						"in": "path",
						"name": "issue_id",
						"required": true,
						"schema": {
							"title": "Issue Id",
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "page",
						"required": true,
						"schema": {
							"title": "Page",
							"type": "integer"
						}
					},
					{
						"description": "Snapped to the next width up",
						"in": "query",
						"name": "w",
						"required": false,
						"schema": {
							"anyOf": [
								{
									"maximum": 100000,
									"minimum": 1,
									"type": "integer"
								},
								{
									"type": "null"
								}
							],
							"description": "Snapped to the next width up",
							"title": "W"
						}
					},
					{
						"description": "Cache buster: the PDF's mtime",
						"in": "query",
						"name": "v",
						"required": false,
						"schema": {
							"anyOf": [
								{
									"type": "string"
								},
								{
									"type": "null"
								}
							],
							"description": "Cache buster: the PDF's mtime",
							"title": "V"
						}
					}
				],
				"responses": {
					"200": {
						"content": {
							"image/webp": {}
						},
						"description": "The rendered page"
					},
					"304": {
						"description": "The client's copy is current"
					},
					"404": {
						"description": "No such issue, or no such page"
					},
					"422": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						},
						"description": "Validation Error"
					}
				},
				"summary": "Page Image",
				"tags": [
					"images"
				]
			}
		},
...
		"/api/issues/{issue_id}/thumb.jpg": {
			"get": {
				"description": "The issue's thumbnail, rendered on demand when the cache has lost it.",
				"operationId": "issue_thumb",
				"parameters": [
					{
						"in": "path",
						"name": "issue_id",
						"required": true,
						"schema": {
							"title": "Issue Id",
							"type": "string"
						}
					},
					{
						"description": "Cache buster: the PDF's mtime",
						"in": "query",
						"name": "v",
						"required": false,
						"schema": {
							"anyOf": [
								{
									"type": "string"
								},
								{
									"type": "null"
								}
							],
							"description": "Cache buster: the PDF's mtime",
							"title": "V"
						}
					}
				],
				"responses": {
					"200": {
						"content": {
							"image/jpeg": {}
						},
						"description": "The image"
					},
					"304": {
						"description": "The client's copy is current"
					},
					"422": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						},
						"description": "Validation Error"
					}
				},
				"summary": "Thumbnail",
				"tags": [
					"images"
				]
			},
			"head": {
				"description": "The issue's thumbnail, rendered on demand when the cache has lost it.",
				"operationId": "issue_thumb_head",
				"parameters": [
					{
						"in": "path",
						"name": "issue_id",
						"required": true,
						"schema": {
							"title": "Issue Id",
							"type": "string"
						}
					},
					{
						"description": "Cache buster: the PDF's mtime",
						"in": "query",
						"name": "v",
						"required": false,
						"schema": {
							"anyOf": [
								{
									"type": "string"
								},
								{
									"type": "null"
								}
							],
							"description": "Cache buster: the PDF's mtime",
							"title": "V"
						}
					}
				],
				"responses": {
					"200": {
						"content": {
							"image/jpeg": {}
						},
						"description": "The image"
					},
					"304": {
						"description": "The client's copy is current"
					},
					"422": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						},
						"description": "Validation Error"
					}
				},
				"summary": "Thumbnail",
				"tags": [
					"images"
				]
			}
		},
...
	}
}